+++ core-bundle/src/EventListener/CsrfTokenCookieListener.php
@@ -107,7 +107,7 @@ class CsrfTokenCookieListener
         $tokens = [];
 
         foreach ($cookies as $key => $value) {
-            if (0 === strncmp($key, $this->cookiePrefix, \strlen($this->cookiePrefix))) {
+            if (0 === strpos($key, $this->cookiePrefix)) {
                 $tokens[substr($key, \strlen($this->cookiePrefix))] = $value;
             }
         }