+++ core-bundle/src/Resources/contao/library/Contao/System.php
@@ -720,7 +720,14 @@ abstract class System
 	public static function getSessionHash($strCookie)
 	{
 		$container = static::getContainer();
-		$strHash = $container->get('session')->getId();
+		$session = $container->get('session');
+
+		if (!$session->isStarted())
+		{
+			$session->start();
+		}
+
+		$strHash = $session->getId();
 
 		if (!$container->getParameter('contao.security.disable_ip_check'))
 		{